+++ das-20200116/alibabacloud_das20200116/models.py
@@ -7380,6 +7380,110 @@ class DescribeHotBigKeysResponseBodyDataBigKeys(TeaModel):
         return self
 
 
+class DescribeHotBigKeysResponseBodyDataHighTrafficKeysHighTrafficKey(TeaModel):
+    def __init__(
+        self,
+        db: int = None,
+        hot: str = None,
+        key: str = None,
+        key_type: str = None,
+        node_id: str = None,
+        size: int = None,
+        in_bytes: int = None,
+        out_bytes: int = None,
+    ):
+        self.db = db
+        self.hot = hot
+        self.key = key
+        self.key_type = key_type
+        self.node_id = node_id
+        self.size = size
+        self.in_bytes = in_bytes
+        self.out_bytes = out_bytes
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.db is not None:
+            result['Db'] = self.db
+        if self.hot is not None:
+            result['Hot'] = self.hot
+        if self.key is not None:
+            result['Key'] = self.key
+        if self.key_type is not None:
+            result['KeyType'] = self.key_type
+        if self.node_id is not None:
+            result['NodeId'] = self.node_id
+        if self.size is not None:
+            result['Size'] = self.size
+        if self.in_bytes is not None:
+            result['inBytes'] = self.in_bytes
+        if self.out_bytes is not None:
+            result['outBytes'] = self.out_bytes
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('Db') is not None:
+            self.db = m.get('Db')
+        if m.get('Hot') is not None:
+            self.hot = m.get('Hot')
+        if m.get('Key') is not None:
+            self.key = m.get('Key')
+        if m.get('KeyType') is not None:
+            self.key_type = m.get('KeyType')
+        if m.get('NodeId') is not None:
+            self.node_id = m.get('NodeId')
+        if m.get('Size') is not None:
+            self.size = m.get('Size')
+        if m.get('inBytes') is not None:
+            self.in_bytes = m.get('inBytes')
+        if m.get('outBytes') is not None:
+            self.out_bytes = m.get('outBytes')
+        return self
+
+
+class DescribeHotBigKeysResponseBodyDataHighTrafficKeys(TeaModel):
+    def __init__(
+        self,
+        high_traffic_key: List[DescribeHotBigKeysResponseBodyDataHighTrafficKeysHighTrafficKey] = None,
+    ):
+        self.high_traffic_key = high_traffic_key
+
+    def validate(self):
+        if self.high_traffic_key:
+            for k in self.high_traffic_key:
+                if k:
+                    k.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        result['HighTrafficKey'] = []
+        if self.high_traffic_key is not None:
+            for k in self.high_traffic_key:
+                result['HighTrafficKey'].append(k.to_map() if k else None)
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        self.high_traffic_key = []
+        if m.get('HighTrafficKey') is not None:
+            for k in m.get('HighTrafficKey'):
+                temp_model = DescribeHotBigKeysResponseBodyDataHighTrafficKeysHighTrafficKey()
+                self.high_traffic_key.append(temp_model.from_map(k))
+        return self
+
+
 class DescribeHotBigKeysResponseBodyDataHotKeysHotKey(TeaModel):
     def __init__(
         self,
@@ -7389,6 +7493,7 @@ class DescribeHotBigKeysResponseBodyDataHotKeysHotKey(TeaModel):
         key_type: str = None,
         lfu: int = None,
         node_id: str = None,
+        size: int = None,
     ):
         # The database in which the key is stored.
         self.db = db
@@ -7402,6 +7507,7 @@ class DescribeHotBigKeysResponseBodyDataHotKeysHotKey(TeaModel):
         self.lfu = lfu
         # The ID of the data shard on the ApsaraDB for Redis instance.
         self.node_id = node_id
+        self.size = size
 
     def validate(self):
         pass
@@ -7424,6 +7530,8 @@ class DescribeHotBigKeysResponseBodyDataHotKeysHotKey(TeaModel):
             result['Lfu'] = self.lfu
         if self.node_id is not None:
             result['NodeId'] = self.node_id
+        if self.size is not None:
+            result['Size'] = self.size
         return result
 
     def from_map(self, m: dict = None):
@@ -7440,6 +7548,8 @@ class DescribeHotBigKeysResponseBodyDataHotKeysHotKey(TeaModel):
             self.lfu = m.get('Lfu')
         if m.get('NodeId') is not None:
             self.node_id = m.get('NodeId')
+        if m.get('Size') is not None:
+            self.size = m.get('Size')
         return self
 
 
@@ -7483,6 +7593,8 @@ class DescribeHotBigKeysResponseBodyData(TeaModel):
         self,
         big_key_msg: str = None,
         big_keys: DescribeHotBigKeysResponseBodyDataBigKeys = None,
+        high_traffic_key_msg: str = None,
+        high_traffic_keys: DescribeHotBigKeysResponseBodyDataHighTrafficKeys = None,
         hot_key_msg: str = None,
         hot_keys: DescribeHotBigKeysResponseBodyDataHotKeys = None,
     ):
@@ -7490,6 +7602,8 @@ class DescribeHotBigKeysResponseBodyData(TeaModel):
         self.big_key_msg = big_key_msg
         # The list of large keys.
         self.big_keys = big_keys
+        self.high_traffic_key_msg = high_traffic_key_msg
+        self.high_traffic_keys = high_traffic_keys
         # The reason why the hot key failed to be queried.
         self.hot_key_msg = hot_key_msg
         # The list of hot keys.
@@ -7498,6 +7612,8 @@ class DescribeHotBigKeysResponseBodyData(TeaModel):
     def validate(self):
         if self.big_keys:
             self.big_keys.validate()
+        if self.high_traffic_keys:
+            self.high_traffic_keys.validate()
         if self.hot_keys:
             self.hot_keys.validate()
 
@@ -7511,6 +7627,10 @@ class DescribeHotBigKeysResponseBodyData(TeaModel):
             result['BigKeyMsg'] = self.big_key_msg
         if self.big_keys is not None:
             result['BigKeys'] = self.big_keys.to_map()
+        if self.high_traffic_key_msg is not None:
+            result['HighTrafficKeyMsg'] = self.high_traffic_key_msg
+        if self.high_traffic_keys is not None:
+            result['HighTrafficKeys'] = self.high_traffic_keys.to_map()
         if self.hot_key_msg is not None:
             result['HotKeyMsg'] = self.hot_key_msg
         if self.hot_keys is not None:
@@ -7524,6 +7644,11 @@ class DescribeHotBigKeysResponseBodyData(TeaModel):
         if m.get('BigKeys') is not None:
             temp_model = DescribeHotBigKeysResponseBodyDataBigKeys()
             self.big_keys = temp_model.from_map(m['BigKeys'])
+        if m.get('HighTrafficKeyMsg') is not None:
+            self.high_traffic_key_msg = m.get('HighTrafficKeyMsg')
+        if m.get('HighTrafficKeys') is not None:
+            temp_model = DescribeHotBigKeysResponseBodyDataHighTrafficKeys()
+            self.high_traffic_keys = temp_model.from_map(m['HighTrafficKeys'])
         if m.get('HotKeyMsg') is not None:
             self.hot_key_msg = m.get('HotKeyMsg')
         if m.get('HotKeys') is not None:
@@ -7674,20 +7799,28 @@ class DescribeHotKeysRequest(TeaModel):
 class DescribeHotKeysResponseBodyDataHotKey(TeaModel):
     def __init__(
         self,
+        category: str = None,
         db: int = None,
         hot: str = None,
+        in_bytes: int = None,
         key: str = None,
         key_type: str = None,
+        node_id: str = None,
+        out_bytes: int = None,
         size: int = None,
     ):
+        self.category = category
         # The database in which the key is stored.
         self.db = db
         # The frequency at which the key is accessed, which indicates the queries per second (QPS) of the key.
         self.hot = hot
+        self.in_bytes = in_bytes
         # The name of the key.
         self.key = key
         # The type of the key.
         self.key_type = key_type
+        self.node_id = node_id
+        self.out_bytes = out_bytes
         # The number of elements in the key.
         self.size = size
 
@@ -7700,28 +7833,44 @@ class DescribeHotKeysResponseBodyDataHotKey(TeaModel):
             return _map
 
         result = dict()
+        if self.category is not None:
+            result['Category'] = self.category
         if self.db is not None:
             result['Db'] = self.db
         if self.hot is not None:
             result['Hot'] = self.hot
+        if self.in_bytes is not None:
+            result['InBytes'] = self.in_bytes
         if self.key is not None:
             result['Key'] = self.key
         if self.key_type is not None:
             result['KeyType'] = self.key_type
+        if self.node_id is not None:
+            result['NodeId'] = self.node_id
+        if self.out_bytes is not None:
+            result['OutBytes'] = self.out_bytes
         if self.size is not None:
             result['Size'] = self.size
         return result
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('Category') is not None:
+            self.category = m.get('Category')
         if m.get('Db') is not None:
             self.db = m.get('Db')
         if m.get('Hot') is not None:
             self.hot = m.get('Hot')
+        if m.get('InBytes') is not None:
+            self.in_bytes = m.get('InBytes')
         if m.get('Key') is not None:
             self.key = m.get('Key')
         if m.get('KeyType') is not None:
             self.key_type = m.get('KeyType')
+        if m.get('NodeId') is not None:
+            self.node_id = m.get('NodeId')
+        if m.get('OutBytes') is not None:
+            self.out_bytes = m.get('OutBytes')
         if m.get('Size') is not None:
             self.size = m.get('Size')
         return self
@@ -13415,17 +13564,22 @@ class DescribeTopHotKeysRequest(TeaModel):
 class DescribeTopHotKeysResponseBodyDataHotKey(TeaModel):
     def __init__(
         self,
+        category: str = None,
         db: int = None,
         hot: str = None,
+        in_bytes: int = None,
         key: str = None,
         key_type: str = None,
         lfu: int = None,
         node_id: str = None,
+        out_bytes: int = None,
     ):
+        self.category = category
         # The database in which the key is stored.
         self.db = db
         # The frequency at which the key is accessed, which indicates the QPS of the key.
         self.hot = hot
+        self.in_bytes = in_bytes
         # The key.
         self.key = key
         # The type of the key.
@@ -13434,6 +13588,7 @@ class DescribeTopHotKeysResponseBodyDataHotKey(TeaModel):
         self.lfu = lfu
         # The ID of the data shard on the ApsaraDB for Redis instance.
         self.node_id = node_id
+        self.out_bytes = out_bytes
 
     def validate(self):
         pass
@@ -13444,10 +13599,14 @@ class DescribeTopHotKeysResponseBodyDataHotKey(TeaModel):
             return _map
 
         result = dict()
+        if self.category is not None:
+            result['Category'] = self.category
         if self.db is not None:
             result['Db'] = self.db
         if self.hot is not None:
             result['Hot'] = self.hot
+        if self.in_bytes is not None:
+            result['InBytes'] = self.in_bytes
         if self.key is not None:
             result['Key'] = self.key
         if self.key_type is not None:
@@ -13456,14 +13615,20 @@ class DescribeTopHotKeysResponseBodyDataHotKey(TeaModel):
             result['Lfu'] = self.lfu
         if self.node_id is not None:
             result['NodeId'] = self.node_id
+        if self.out_bytes is not None:
+            result['OutBytes'] = self.out_bytes
         return result
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('Category') is not None:
+            self.category = m.get('Category')
         if m.get('Db') is not None:
             self.db = m.get('Db')
         if m.get('Hot') is not None:
             self.hot = m.get('Hot')
+        if m.get('InBytes') is not None:
+            self.in_bytes = m.get('InBytes')
         if m.get('Key') is not None:
             self.key = m.get('Key')
         if m.get('KeyType') is not None:
@@ -13472,6 +13637,8 @@ class DescribeTopHotKeysResponseBodyDataHotKey(TeaModel):
             self.lfu = m.get('Lfu')
         if m.get('NodeId') is not None:
             self.node_id = m.get('NodeId')
+        if m.get('OutBytes') is not None:
+            self.out_bytes = m.get('OutBytes')
         return self